pyproject.toml
[project]
name = "nonebot-plugin-kawaii-robot"
dynamic = ["version"]
description = "使用 Kyomotoi/AnimeThesaurus 的 NoneBot2 的回复（文i）插件"
authors = [
    { name = "karisaya", email = "1048827424@qq.com" },
    { name = "LgCookie", email = "lgc2333@126.com" },
]
dependencies = [
    "nonebot2>=2.4.2",
    "nonebot-plugin-alconna>=0.59.0",
    "nonebot-plugin-userinfo>=0.2.6",
    "nonebot-plugin-session>=0.3.2",
    "cookit[pydantic]>=0.12.0.post1",
    "anyio>=4.9.0",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "AGPL-3.0" }

[project.urls]
homepage = "https://github.com/KarisAya/nonebot_plugin_kawaii_robot/"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_kawaii_robot/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
